khetibandhu_backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
fuzzywuzzy==0.18.0